chore(docker-build-push.yaml): add step to output repository name and fix formatting of images input
The commit adds a new step to the workflow that outputs the repository
name. This step is useful for debugging and verifying the correct
repository is being used. Additionally, the formatting of the images
input in the setup docker metadata step is fixed to ensure proper
parsing of the repository name.