chore(golang-lint.yml): update Go version and linter version
The Go version has been updated from '~1.20.6' to '~1.21.5' and the
linter version from 'v1.52.2' to 'v1.55.2'. These updates ensure that
the project is using the latest stable versions of Go and the linter,
which include new features, performance improvements, and bug fixes.
This helps to maintain the quality and reliability of the codebase.