feat(docker-build-push-pr.yaml): disable 'latest' tag in Docker metadata action
The 'latest' tag is now disabled in the Docker metadata action. This
change is made to prevent the 'latest' tag from being automatically
assigned to new Docker images, which can help avoid potential issues in
production environments where the 'latest' tag might be mistakenly
pulled instead of a specific version.