chore(docker-build-push-pr.yaml): comment out tags section in docker build and push action
The tags section in the docker build and push action has been commented
out. This is done to prevent automatic tagging of docker images during
the build and push process. This change allows for more control over the
tagging process, ensuring that tags are only assigned when necessary and
appropriate.