feat(docker-build-push-pr.yaml): enable fetch_depth and uncomment tags in docker build and push action
The fetch_depth is set to 0 to fetch all history for all branches and
tags which is useful for generating a complete changelog. The tags for
the docker build and push action are uncommented to enable tagging of
docker images with the type of git reference and the event type, as well
as the short format of the git SHA. This provides more information about
the docker image and its corresponding git commit.