chore(deploy-app.yaml): refactor GitHub Actions workflow to use environment variables for module paths
This change introduces environment variables `IMAGE_MODULE` and
`KOPS_MODULE` to store the paths of the modules used in the GitHub
Actions workflow. This makes the workflow script cleaner and easier to
maintain by reducing repetition and improving readability. Additionally,
the script for building and pushing the container image and exporting
kubectl configuration has been simplified to use these variables, making
the workflow more efficient and easier to update in the future.