chore(deploy-backend.yaml): streamline GitHub Actions workflow by removing unused inputs and comments
This change simplifies the GitHub Actions workflow for deploying the
backend by removing the 'deploy_group' input and several commented-out
sections that were not being used. This makes the workflow easier to
understand and maintain. Additionally, the workflow now includes steps
for checking out the repository and authenticating with Google Cloud,
which are necessary for the deployment process. The 'verify pulumi' step
has been uncommented and updated to use these new authentication
details, ensuring that Pulumi can manage infrastructure as expected.