feat(deploy-backend.yaml): add inputs for app and image_tag to specify deployment details
The deployment workflow now requires two new inputs: 'app' and
'image_tag', which specify the backend application and its Docker image
tag to be deployed. This change allows for more dynamic and specific
deployments per application. Additionally, the step to download kubectl
has been updated to clarify that it is downloaded from an artifact, and
a new deployment step has been added to utilize the Pulumi module for
deploying the specified backend application.