feat(.github/workflows): add default value and input handling for artifact_name
The `artifact_name` input parameter in the GitHub Actions workflows now
has a default value of `kubectl-config` in `composite-deploy.yaml`,
making it optional to specify this when the workflow is triggered.
Additionally, the handling of `artifact_name` is added to the jobs in
`composite-deploy.yaml` to ensure that the artifact operations are
dynamic based on the input or default value.
In `deploy-setup.yaml`, the `artifact_name` input parameter is
explicitly required and used to specify the name of the artifact in the
upload step, ensuring that the artifact name is configurable and can be
specified at the workflow dispatch level.