feat(composite-deploy.yaml): modify repository and conditional deployment logic
The default application type in the GitHub Actions workflow is updated
from 'microservice' to 'backend' to better reflect the current
deployment strategies and terminologies used within the project.
Additionally, the repository for checking out code is changed to
'dictybase-docker/dagger-of-dcr' to align with the new source of
deployment scripts. The deployment condition is also expanded to include
'frontend', allowing the same deployment step to handle multiple types
of applications, increasing the flexibility of the workflow.