ci(golang-lint.yml): update workflow to use justfile for linting
Switches the checkout action to version 4 and specifies a different
repository and branch to align with the new project structure.
Replaces the Go setup and golangci-lint actions with a setup for the
just binary, which is now used to run the linter. This change
simplifies the workflow by consolidating the linting process into a
single command, improving maintainability and reducing complexity.