ci(golang-lint.yaml): add version input for golanglint-ci to allow specifying the linter version
Introduce a new input parameter 'version' to the GitHub Actions
workflow for golang-lint. This allows users to specify the version
of golanglint-ci they wish to use, providing more control over
the linting process. The default version is set to "v1.62.2-alpine"
to ensure consistency and stability in the linting environment.